+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Fix markdown max image width (front) [#1707](https://github.com/opendatateam/udata/pull/1707)
 - Adds ETag to internal avatar for efficient caching control [#1712](https://github.com/opendatateam/udata/pull/1712)
 - Show resource type in modal (front) [#1714](https://github.com/opendatateam/udata/pull/1714)
+- Starts using harvest backend `config` (validation, API exposition, `HarvestFilters`...) [#1716](https://github.com/opendatateam/udata/pull/1716)
 
 ## 1.3.12 (2018-05-31)